[Redux] Rename queriesSlice to transactionsSlice (#5598)
* Move getAccountsById to accountsSlice

* [autofix.ci] apply automated fixes

* [Redux] Rename queriesSlice to transactionsSlice

* Fix import

* Release notes

* Remove nonexistent saved state

* [autofix.ci] apply automated fixes

---------
