+++ upcoming-release-notes/5598.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [joel-jeremy]
+---
+
+[Redux] Rename queriesSlice to transactionsSlice
